# apps/docs/static/openapi/framework-live.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/notifications/{id}": {
      "get": {
        "operationId": "NotificationsController_getNotification",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Notification identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": false,
            "in": "path",
            "description": "Optional locale code used to localize response values.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns notification details."
          }
        },
        "summary": "Get notification by id",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/notifications": {
      "get": {
        "operationId": "NotificationsController_getNotificationList",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Notification type filter.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "description": "Notification priority filter.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Notification status filter.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "description": "Date range start as ISO string.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "description": "Date range end as ISO string.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Sort expression from query string, for example `createdAt_DESC`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "description": "Optional locale code used to localize response values.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns notifications list."
          }
        },
        "summary": "List notifications",
        "tags": [
          "notifications"
        ]
      },
      "post": {
        "operationId": "NotificationsController_markNotificationAs",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkNotificationAsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification status updated."
          }
        },
        "summary": "Update notification status",
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/api/users/me/customers/{id}": {
      "get": {
        "operationId": "UserController_getCustomerForCurrentUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Customer identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns customer details for current user."
          }
        },
        "summary": "Get current user customer by id",
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/invoices": {
      "get": {
        "operationId": "InvoiceController_getInvoiceList",
        "parameters": [
          {
            "name": "paymentStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns invoices list."
          }
        },
        "summary": "List invoices",
        "tags": [
          "invoices"
        ]
      }
    },
    "/api/invoices/{id}": {
      "get": {
        "operationId": "InvoiceController_getInvoice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns invoice details."
          }
        },
        "summary": "Get invoice by id",
        "tags": [
          "invoices"
        ]
      }
    },
    "/api/invoices/{id}/pdf": {
      "get": {
        "operationId": "InvoiceController_getInvoicePdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns invoice PDF file."
          }
        },
        "summary": "Download invoice PDF",
        "tags": [
          "invoices"
        ]
      }
    },
...
    "/api/customers/addresses": {
      "get": {
        "operationId": "CustomersController_getAddresses",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns customer addresses list."
          }
        },
        "summary": "List customer addresses",
        "tags": [
          "customers"
        ]
      },
      "post": {
        "operationId": "CustomersController_createAddress",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddressBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer address created."
          }
        },
        "summary": "Create customer address",
        "tags": [
          "customers"
        ]
      }
    },
    "/api/customers/addresses/{id}": {
      "get": {
        "operationId": "CustomersController_getAddress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Customer address identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns customer address details."
          }
        },
        "summary": "Get customer address by id",
        "tags": [
          "customers"
        ]
      },
      "patch": {
        "operationId": "CustomersController_updateAddress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Customer address identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddressBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer address updated."
          }
        },
        "summary": "Update customer address",
        "tags": [
          "customers"
        ]
      },
      "delete": {
        "operationId": "CustomersController_deleteAddress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Customer address identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer address deleted."
          }
        },
        "summary": "Delete customer address",
        "tags": [
          "customers"
        ]
      }
    },
    "/api/customers/addresses/{id}/default": {
      "post": {
        "operationId": "CustomersController_setDefaultAddress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Customer address identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default customer address updated."
          }
        },
        "summary": "Set default customer address",
        "tags": [
          "customers"
        ]
      }
    },
...
  },
  "info": {
    "title": "O2S API Harmonization HTTP API",
    "description": "Live API documentation generated from current NestJS controllers and DTOs.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "MarkNotificationAsRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Notification identifier."
          },
          "status": {
            "type": "string",
            "description": "Target notification status to set."
          }
        },
        "required": [
          "id",
          "status"
        ]
      },
...
      "Address": {
        "type": "object",
        "properties": {}
      },
      "CreateAddressBody": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "description": "Optional address label visible in UI (for example `Home`, `Office`)."
          },
          "isDefault": {
            "type": "boolean",
            "description": "When true, sets the new address as default."
          },
          "address": {
            "description": "Full address payload.",
            "allOf": [
              {
                "$ref": "#/components/schemas/Address"
              }
            ]
          }
        },
        "required": [
          "address"
        ]
      },
      "UpdateAddressBody": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "description": "Optional address label visible in UI."
          },
          "isDefault": {
            "type": "boolean",
            "description": "When true, sets the updated address as default."
          },
          "address": {
            "description": "Updated full address payload.",
            "allOf": [
              {
                "$ref": "#/components/schemas/Address"
              }
            ]
          }
        }
      },
...
    }
  }
}