# packages/framework/src/modules/invoices/invoices.controller.ts
import * as Invoices from '.';
import { Controller, Get, Headers, Param, Query, Res, UseInterceptors } from '@nestjs/common';
import { ApiOperation, ApiParam, ApiQuery, ApiResponse, ApiTags } from '@nestjs/swagger';
import type { Response } from 'express';
import { Observable, map } from 'rxjs';

import { LoggerService } from '@o2s/utils.logger';

import { GetInvoiceListQuery, GetInvoiceParams } from './invoices.request';
import { InvoiceService } from './invoices.service';
import { AppHeaders } from '@/utils/models/headers';

/**
 * HTTP controller for invoices. Base path: `/invoices`. All methods delegate to {@link InvoiceService}.
 */
@Controller('/invoices')
@UseInterceptors(LoggerService)
@ApiTags('invoices')
export class InvoiceController {
    constructor(protected readonly invoiceService: InvoiceService) {}

    @Get()
    @ApiOperation({ summary: 'List invoices' })
    @ApiQuery({
        name: 'query',
        required: false,
        type: String,
        description: 'Invoice list filters and pagination query.',
    })
    @ApiResponse({ status: 200, description: 'Returns invoices list.' })
    getInvoiceList(
        @Query() query: GetInvoiceListQuery,
        @Headers() headers: AppHeaders,
    ): Observable<Invoices.Model.Invoices> {
        return this.invoiceService.getInvoiceList(query, headers.authorization);
    }

    @Get(':id')
    @ApiOperation({ summary: 'Get invoice by id' })
    @ApiParam({ name: 'id', type: String, description: 'Invoice identifier.' })
    @ApiResponse({ status: 200, description: 'Returns invoice details.' })
    getInvoice(@Param() params: GetInvoiceParams, @Headers() headers: AppHeaders): Observable<Invoices.Model.Invoice> {
        return this.invoiceService.getInvoice(params, headers.authorization);
    }

    @Get(':id/pdf')
    @ApiOperation({ summary: 'Download invoice PDF' })
    @ApiParam({ name: 'id', type: String, description: 'Invoice identifier.' })
    @ApiResponse({ status: 200, description: 'Returns invoice PDF file.' })
    getInvoicePdf(
        @Param() params: GetInvoiceParams,
        @Headers() headers: AppHeaders,
        @Res() res: Response,
    ): Observable<void> {
        return this.invoiceService.getInvoicePdf(params, headers.authorization).pipe(
            map((pdf) => {
                res.setHeader('Content-Type', 'application/pdf');
                res.setHeader('Content-Disposition', `attachment; filename="invoice-${params.id}.pdf"`);
                res.setHeader('Content-Length', pdf.byteLength);
                res.end(pdf);
            }),
        );
    }
}